# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Нууу очень крутая апишка",
        "description": "Документация для REST API Laravel",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
        "/api/v1/auth/registration": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "628a0714e9a4ed846ed1a311a6076eeb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь успешно зарегестрирован"
                    },
                    "422": {
                        "description": "Неверные данные"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "9d0bb5c770cbec19b0b9566606ef4adb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный вход"
                    },
                    "401": {
                        "description": "Неверные учетные данные"
                    },
                    "422": {
                        "description": "Неверные данные"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выход пользователя",
                "description": "Отзывает токен текущего пользователя и завершает сессию авторизации.",
                "operationId": "9c11fea5b676e17442cb348406aefe9e",
                "responses": {
                    "200": {
                        "description": "Выход выполнен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Выход выполнен успешно"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Получить информацию о текущем пользователе",
                "description": "Возвращает данные авторизованного пользователя по токену.",
                "operationId": "7edb820901b8b3a880fc4d7a0a7290fb",
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Иван Иванов"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "user@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-10-21T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/email-verify/{user}": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Отправить ссылку для подтверждения email",
                "description": "Создает токен и возвращает ссылку для подтверждения email указанного пользователя.",
                "operationId": "6a114565532c0f859a327041df493250",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ссылка создана успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ссылка для подтверждения email создана."
                                        },
                                        "verify_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/auth/email-verify?user=1&token=abc123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email уже подтверждён"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        },
        "/api/v1/auth/email-verify": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Подтвердить email по токену",
                "description": "Проверяет токен подтверждения email и устанавливает дату верификации.",
                "operationId": "e40dc331738bbf2df00d8b621dc2d0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user",
                                    "token"
                                ],
                                "properties": {
                                    "user": {
                                        "description": "ID пользователя",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "token": {
                                        "description": "Токен подтверждения email",
                                        "type": "string",
                                        "example": "abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email успешно подтверждён"
                    },
                    "400": {
                        "description": "Неверный или истёкший токен"
                    },
                    "404": {
                        "description": "Запрос подтверждения не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
...
    },
...
}